Updated m1_electron_beam_model
Set subroutine definition withint the 80 characters limit.

Added alternative longer name but kept it commented.

Reorganized unit conversions so that they are all on the same place.

Renamed and reorganized mu and sigma values for the positions and
energy distributions.

Renamed the electron_energy_1 and 2 variables to just energy_1 and 2.

Defined translation_vector with a literal name now, included the use
of a vector_translation subroutine instead of doing the translation
directly in the code.

Fixed tranformation operations order.