Testing solution for personalized indexes with negative values
The fix consists in setting the lower boundary when declaring the arrays,
so that it started with the negative position, using the material_boundaries array.

The test consists in printing the lower and upper bounds of the arrays both in main
and inside subroutines. The values of the arrays are also read and printed to force
an error, if still present.

The fix and tests were applied to time_step and compute_trajectory subroutines.

Included as comments: dev way of calculating acceleration due to atom, position and
acceleration update in time_step. These will be tested in case the results between
dev and src are inconsistent, even after fixing the personalized indexes bug in both.